Update tyeglass_assignment-4
I added logic that checks if the user’s study choice is in the list of study options and adjusts study hours and stress level differently depending on the subject. I also included a final check that prints an error message if the user enters a choice not in the list.